Refactor approve button container
