+++ src/cow-react/modules/swap/hooks/useSwapButtonContext.ts
@@ -15,7 +15,7 @@ import { swapFlow } from 'cow-react/modules/swap/services/swapFlow'
 import { useApproveCallbackFromTrade } from 'hooks/useApproveCallback'
 import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
 import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
-import { ApproveButtonProps } from 'cow-react/modules/swap/containers/ApproveButton'
+import { ApproveButtonProps } from '@src/cow-react/modules/swap/containers/ApproveErrorButtons'
 import { getSwapButtonState } from 'cow-react/modules/swap/helpers/getSwapButtonState'
 import { SwapButtonContext } from 'cow-react/modules/swap/pure/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
